requirements.txt
piper-tts
onnxruntime          # или onnxruntime-gpu
numpy
wave

